services/auth_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
redis==5.0.4
httpx==0.27.0
structlog==24.1.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
pyotp==2.9.0
cryptography==42.0.5
aiosqlite==0.20.0
pytest==8.2.0
pytest-asyncio==0.23.6
anyio==4.3.0
